+++ src/pkmodelling/pkrun2.cpp
@@ -22,10 +22,10 @@ Pkrun2::Pkrun2(std::vector<double> & tt1, std::vector< std::vector<double> > & y
 
     // Get dimensions
     n_par = 4; // number of parameters in model function f
-    m_t1 = t1.size(); // number of data pairs
-    mrows=y1.size();
-    ncols=y1[0].size();
-    n_t101= T101.size();
+    m_t1 = (int)t1.size(); // number of data pairs
+    mrows = (int)y1.size();
+    ncols = (int)y1[0].size();
+    n_t101 = (int)T101.size();
 
     // Assign memory to variable arrays and populate temporal array for a single instance
     t = new double [m_t1];
